code/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25
pandas==2.1.4
plotly==5.18.0
pymysql==1.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
cryptography==44.0.1
pyspark==3.5.0
pyarrow==14.0.1
